requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=requirements.txt --pip-args=None requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.10.0
    # via
    #   channels
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.2
    # via celery
celery[redis]==5.5.3
    # via -r requirements.in
certifi==2025.11.12
    # via sentry-sdk
cffi==2.0.0
    # via cryptography
channels==4.3.1
    # via -r requirements.in
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via -r requirements.in
dj-rest-auth==7.0.1
    # via -r requirements.in
django==5.2.7
    # via
    #   -r requirements.in
    #   channels
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   django-filter
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==65.13.0
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.29.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kombu[redis]==5.5.4
    # via celery
packaging==25.0
    # via kombu
prompt-toolkit==3.0.52
    # via click-repl
psycopg[binary]==3.2.12
    # via -r requirements.in
psycopg-binary==3.2.12
    # via psycopg
pycparser==2.23
    # via cffi
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via celery
pyyaml==6.0.3
    # via drf-spectacular
redis==5.2.1
    # via
    #   django-redis
    #   kombu
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
sentry-sdk==2.44.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via
    #   psycopg
    #   referencing
tzdata==2025.2
    # via
    #   django
    #   kombu
    #   psycopg
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
